rearrange some stuff, fix unsound POD impl
